WRDN-499 Cache the rota & locations when the warden login to the app

+++ lib/models/location.dart
@@ -171,7 +171,7 @@ LocationWithZones _$LocationWithZonesFromJson(Map<String, dynamic> json) {
     Longitude: json['Longitude'],
     Latitude: json['Latitude'],
     Notes: json['Notes'],
-    Distance: json['Distance'],
+    Distance: json['Distance'].toDouble(),
     Zones: zonesList,
     operationalPeriodHistories: operationalPeriodsList,
     UpperAmount: json['Rates'] != null
